Refactor OpenApiMerger into OpenApiDocument extension methods
